\name{qcondbvtcop}
\alias{qcondbvtcop}
\title{C_[2|1]^[-1](p|u) for bivariate Student t copula}
\description{
bivariate Student t copula conditional quantile
}

\usage{
qcondbvtcop(p,u,cpar)
}
\arguments{
\item{p}{0<p<1, could be a vector}
\item{u}{0<u<1, could be a vector }
\item{cpar}{copula parameter: 2-vector with -1<rho<1, df>0}
}
\value{
 conditional quantiles of bivariate Student t copula}
