\name{rmvn}
\alias{rmvn}
\title{Random multivariate normal (standard N(0,1) margins)}
\description{
Random multivariate normal (standard N(0,1) margins)
}

\usage{
rmvn(n,rmat)
}
\arguments{
\item{n}{simulation sample size}
\item{rmat}{correlation matrix }
}
\value{
 nxd matrix, where d=nrow(rmat)}
