\name{rmvt}
\alias{rmvt}
\title{Random multivariate t (standard t(nu) margins)}
\description{
Random multivariate t (standard t(nu) margins)
}

\usage{
rmvt(n,rmat,nu)
}
\arguments{
\item{n}{simulation sample size}
\item{rmat}{correlation matrix }
\item{nu}{degree of freedom parameter }
}
\value{
 nxd matrix, where d=nrow(rmat)}
