\name{RVtrunc2cor}
\alias{RVtrunc2cor}
\title{compute correlation matrix from 2-truncated R-vine}
\description{
compute correlation matrix from 2-truncated R-vine
}

\usage{
RVtrunc2cor(RVobj)
}
\arguments{
\item{RVobj}{R-vine object with vine array and partial correlation matrix variable 1 is the latent variable; list with $Varray (d+1)x(d+1), $PCor 2x(d+1)}
}
\value{
 correlation matrix for 2-truncated vine structure based on partial correlations in tree 2
}
\details{
not exported
}
